requirements.txt
torch>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.10.0
